Added paginated objects display, threads and stream() sorting.
App has now a new class, named Paginator, that allows the UI to display
the content of the repos in pages. Manager uses threads and stream() to
get the info and sort it before passing it over to the paginator. TO DO:
add filtering and a way to choose sorting order in the manager, also to
refactor a little bit of the code.